+++ cmd/deja/page_titles_count_test.go
@@ -93,7 +93,7 @@ func countWordForTest(n int) string {
 	words := []string{"zero", "one", "two", "three", "four", "five", "six", "seven",
 		"eight", "nine", "ten", "eleven", "twelve", "thirteen", "fourteen", "fifteen",
 		"sixteen", "seventeen", "eighteen", "nineteen", "twenty", "twenty-one", "twenty-two",
-		"twenty-three", "twenty-four", "twenty-five", "twenty-six"}
+		"twenty-three", "twenty-four", "twenty-five", "twenty-six", "twenty-seven"}
 	if n < 0 || n >= len(words) {
 		return fmt.Sprint(n)
 	}